REFACTOR Renaming package to match new name of interface
