+++ PuzzleSolverProject/PuzzleSolverProject.csproj
@@ -58,20 +58,20 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="GetNeighbors\DirectionEnum.cs" />
-    <Compile Include="GetNeighbors\DownLeftDirectionSearchStrategy.cs" />
-    <Compile Include="GetNeighbors\DownRightDirectionSearchStrategy.cs" />
-    <Compile Include="GetNeighbors\DirectionSearchFactory.cs" />
-    <Compile Include="GetNeighbors\IDirectionSearchStrategy.cs" />
-    <Compile Include="GetNeighbors\DownDirectionSearchStrategy.cs" />
-    <Compile Include="GetNeighbors\LeftDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\DirectionEnum.cs" />
+    <Compile Include="DirectionSearchStrategies\DownLeftDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\DownRightDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\DirectionSearchFactory.cs" />
+    <Compile Include="DirectionSearchStrategies\IDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\DownDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\LeftDirectionSearchStrategy.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PuzzleFileParser.cs" />
-    <Compile Include="GetNeighbors\RightDirectionSearchStrategy.cs" />
-    <Compile Include="GetNeighbors\UpDirectionSearchStrategy.cs" />
-    <Compile Include="GetNeighbors\UpLeftDirectionSearchStrategy.cs" />
-    <Compile Include="GetNeighbors\UpRightDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\RightDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\UpDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\UpLeftDirectionSearchStrategy.cs" />
+    <Compile Include="DirectionSearchStrategies\UpRightDirectionSearchStrategy.cs" />
     <Compile Include="WordsMapDecorator.cs" />
     <Compile Include="WordSeachAlgorithm.cs" />
     <Compile Include="WordSearchPuzzle.cs" />
