+++ PuzzleSolverProject/WordSeachAlgorithm.cs
@@ -1,4 +1,4 @@
-﻿using PuzzleSolverProject.GetNeighbors;
+﻿using PuzzleSolverProject.DirectionSearchStrategies;
 using System;
 using System.Collections.Generic;
 using System.Linq;
